Move core.extensions to seperate module
