+++ saleor/celeryconf.py
@@ -3,7 +3,7 @@ import os
 from celery import Celery
 from django.conf import settings
 
-from .core.extensions import discover_plugins_modules
+from .extensions import discover_plugins_modules
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "saleor.settings")
