+++ saleor/core/__init__.py
@@ -1,8 +1,6 @@
 from django.conf import settings
 from django.core.checks import Warning, register
 
-from .extensions.checks import check_extensions  # NOQA: F401
-
 TOKEN_PATTERN = (
     "(?P<token>[0-9a-z]{8}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{4}" "-[0-9a-z]{12})"
 )
