+++ saleor/dashboard/taxes/views.py
@@ -13,11 +13,11 @@ from django.views.decorators.http import require_POST
 from django_countries.fields import Country
 from django_prices_vatlayer.models import VAT
 
-from ...core.extensions.plugins.vatlayer import TaxRateType, get_taxes_for_country
 from ...core.utils import get_paginator_items
 from ...dashboard.taxes.filters import TaxFilter
 from ...dashboard.taxes.forms import TaxesConfigurationForm
 from ...dashboard.views import staff_member_required
+from ...extensions.plugins.vatlayer import TaxRateType, get_taxes_for_country
 
 logger = logging.getLogger(__name__)
