+++ saleor/order/utils.py
@@ -8,11 +8,11 @@ from prices import Money, TaxedMoney
 
 from ..account.utils import store_user_address
 from ..checkout import AddressType
-from ..core.extensions.manager import get_extensions_manager
 from ..core.taxes import zero_money
 from ..core.weight import zero_weight
 from ..dashboard.order.utils import get_voucher_discount_for_order
 from ..discount.models import NotApplicable
+from ..extensions.manager import get_extensions_manager
 from ..order import FulfillmentStatus, OrderStatus, emails
 from ..order.models import Fulfillment, FulfillmentLine, Order, OrderLine
 from ..payment import ChargeStatus
